==> analysis_for_original_data/0_Background.R <==



#Reproduces the single-cell RNA-seq processing, trajectory analysis, differential expression and visualisations used in
#"αTIGIT-IL2 achieves tumor regression by promoting tumor-infiltrating regulatory T-cell fragility (submmited)"
# The figures used in article will be specified annotated when they are generated in codes. 


# The raw data can be downloaded from the GSA database (https://ngdc.cncb.ac.cn/gsa/) with the accession number: CRA019549
# or contact the authors ( ustczxh@ustc.edu.cn ) for the raw data

# Attention: The code works after getting the filtered data from the raw data by 10X Genomics Cell Ranger pipeline
# One need to run the Cell Ranger pipeline to get the filtered data first, then run the code below
